Toimii jotenkin ja lisäsin accountid tietokantaan, että voi tarkistaa onko oikeutta päivittää tai poistaa postausta.

+++ backend/models/groupPosts.js
@@ -3,9 +3,9 @@
 import { pool } from "../helpers/dbHelper.js"
 
 // TODO: Add Finnkino times to DB or provide parameters for Finnkino API to save FinnkinoTimes
-const queryPostGroupPost = async (postid, groupid, posttext, movieid) => {
-    return await pool.query(`INSERT INTO "groupposts" (postid, fk_groupid, posttext, movieid) VALUES ($1, $2, $3, $4) RETURNING *`,
-        [postid, groupid, posttext, movieid]
+const queryPostGroupPost = async (postid, groupid, posttext, movieid, accountid) => {
+    return await pool.query(`INSERT INTO "groupposts" (postid, fk_groupid, posttext, movieid, fk_accountid) VALUES ($1, $2, $3, $4, $5) RETURNING *`,
+        [postid, groupid, posttext, movieid, accountid]
     )
 }
 
@@ -21,15 +21,15 @@ const queryGroupPostsByPostId = async (postId) => {
 }
 
 
-const queryUpdateGroupPost = async (postid, groupid, posttext, movieid)=> {
+const queryUpdateGroupPost = async (postid, groupid, posttext, movieid, accountid)=> {
     return await pool.query(`UPDATE "groupposts"
-SET postid = $1, groupid = $2, posttext = $3, movieid = $4
-WHERE postid = $1 RETURNING *`, [postid, groupid, posttext, movieid]
+SET postid = $1, fk_groupid = $2, posttext = $3, movieid = $4
+WHERE postid = $1 and fk_accountid = $5 RETURNING *`, [postid, groupid, posttext, movieid, accountid]
     )
 }
 
-const queryDeleteGroupPost = async (postid) => {
-    return await pool.query(`DELETE FROM "groupposts" WHERE postid=$1 RETURNING *`, [postid])
+const queryDeleteGroupPost = async (postid, accountid) => {
+    return await pool.query(`DELETE FROM "groupposts" WHERE postid=$1 and fk_accountid = $2 RETURNING *`, [postid, accountid])
 }
 
 export { queryPostGroupPost,
